requirement.txt
Django==2.2.13
django-import-export==2.2.0
djangorestframework==3.11.0
djongo==1.3.2
dnspython==1.16.0
et-xmlfile==1.0.1
pymongo==3.10.1
requests==2.24.0

